majdnem mukodo tabla
