+++ client/src/App.js
@@ -1,44 +1,19 @@
-import { useEffect, useState } from "react";
 import "./App.css";
+import { useEffect, useState } from "react";
+import SearchBar from "./components/SearchBar";
+import AvailabilityContainer from "./components/AvailabilityContainer";
 
 import axios from "axios";
 
 const App = () => {
-  const url = "http://localhost:5000/api/games/";
-  const [talalat, setTalalat] = useState(null);
-  const [term, setTerm] = useState();
+    const [chosen, setChosen] = useState(null);
 
-  useEffect(() => {
-    const response = axios
-      .get(`http://localhost:5000/api/games/${term}`)
-      .then((res) => {
-        console.log(res.data);
-        setTalalat(res.data);
-        talalat.map((item) => {
-          console.log(item.name);
-        });
-      })
-      .catch((err) => console.error(err));
-  }, [term]);
-
-  const searchInput = (e) => {
-    //console.log(e.target.value);
-    setTerm(e.target.value);
-  };
-
-  return (
-    <div className="App">
-      <h1>Hello</h1>
-      <input onChange={searchInput} type="text" />
-      {talalat && (
-        <ul>
-          {talalat.map((item) => (
-            <li>{item.name}</li>
-          ))}
-        </ul>
-      )}
-    </div>
-  );
+    return (
+        <div className="App">
+            <SearchBar setChosen={setChosen} />
+            {chosen && <AvailabilityContainer chosen={chosen} />}
+        </div>
+    );
 };
 
 export default App;
